Migrate from Portage:  - speed up significance pruning, by optimizing phrasepair-contingency,    speeding up about tenfold.  - reduce the memory requirement for creating TPSAs (and thus for significance    pruning) byu a factor of about two.  - rewrite phrasepair-contingency to use the new TSAs classes, rather than the    old SUFA classes.  - migrate the TSA classes, including mmtsa.find, with its new sampling options.    (the option to create prefix arrays is not migrated yet, however)  - Tokenized text uses 1 or more tab or space characters as separator - respect    that rule in the tpt module too, in particular in vocab.build.cc and    tpt_tokenindex.cc, both used in creating TPSAs.  - Fix a bug in textldm2tpldm.sh, which made it not work correctly.

+++ src/tpt/mmctrack.build.cc
@@ -175,7 +175,7 @@ int MAIN(argc, argv)
     numwrite(out,index[i]);
 
   idxSize = index.size();
-  if (size_t(index.size()) != size_t(idxSize))
+  if (uint64_t(index.size()) != uint64_t(idxSize))
     cerr << efatal << "overflow writing corpus track file " << ctrackFile
          << ": too many sentences for index size type.  sizeof(index size type)="
          << sizeof(idxSize) << exit_1;
